Modularize sandbox support
Sandboxes can now be created by simply making a new python file and adding their config info into settings.
