+++ postprocess/post_b64.py
@@ -1,9 +1,8 @@
-import io
 import re
 import hashlib
+import importlib
 import gzip
 import logging
-import requests
 from base64 import b64decode
 # This gets the raw paste and the paste_data json object
 from common import parse_config
@@ -45,6 +44,7 @@ def run(results, raw_paste_data, paste_object):
                     paste_object["decompressed_stream"] = encoded
                 except Exception as e:
                     logger.error("Unable to decompress gzip stream")
+
             if rule == 'b64_exe':
                 try:
                     raw_exe = b64decode(raw_paste_data)
@@ -55,47 +55,18 @@ def run(results, raw_paste_data, paste_object):
                     # We are guessing that the sample has been submitted, and crafting a URL
                     paste_object["VT"] = 'https://www.virustotal.com/#/file/{0}'.format(paste_object["exe_md5"])
 
-                    # Cuckoo
-                    if conf["post_process"]["post_b64"]["cuckoo"]["enabled"]:
-                        logger.info("Submitting to Cuckoo")
-                        try:
-                            task_id = send_to_cuckoo(raw_exe, paste_object["pasteid"])
-                            paste_object["Cuckoo Task ID"] = task_id
-                            logger.info("exe submitted to Cuckoo with task id {0}".format(task_id))
-                        except Exception as e:
-                            logger.error("Unabled to submit sample to cuckoo")
-
-                    # Viper
-                    if conf["post_process"]["post_b64"]["viper"]["enabled"]:
-                        send_to_cuckoo(raw_exe, paste_object["pasteid"])
-
-                    # VirusTotal
+                    # If sandbox modules are enabled then submit the file
+                    for sandbox, sandbox_values in conf["sandboxes"].items():
+                        if sandbox_values["enabled"]:
+                            logger.info("Uploading file {0} using {1}".format(paste_object["pasteid"], sandbox_values["module"]))
+                            sandbox_module = importlib.import_module(sandbox_values["module"])
+                            paste_object = sandbox_module.upload_file(raw_exe, paste_object)
 
                 except Exception as e:
                     logger.error("Unable to decode exe file")
-
 
     # Get unique domain count
     # Update the json
 
     # Send the updated json back
     return paste_object
-
-
-def send_to_cuckoo(raw_exe, pasteid):
-    cuckoo_ip = conf["post_process"]["post_b64"]["cuckoo"]["api_host"]
-    cuckoo_port = conf["post_process"]["post_b64"]["cuckoo"]["api_port"]
-    cuckoo_host = 'http://{0}:{1}'.format(cuckoo_ip, cuckoo_port)
-    submit_file_url = '{0}/tasks/create/file'.format(cuckoo_host)
-    files = {'file': ('{0}.exe'.format(pasteid), io.BytesIO(raw_exe))}
-    submit_file = requests.post(submit_file_url, files=files).json()
-    task_id = None
-    try:
-        task_id = submit_file['task_id']
-    except KeyError:
-        try:
-            task_id = submit_file['task_ids'][0]
-        except KeyError:
-            logger.error(submit_file)
-
-    return task_id
